Overhaul panels applications and update guest checklist for 2025

+++ uber/site_sections/panels_admin.py
@@ -12,25 +12,34 @@ from uber.config import c
 from uber.decorators import ajax, all_renderable, csrf_protected, csv_file
 from uber.errors import HTTPRedirect
 from uber.models import AssignedPanelist, Attendee, AutomatedEmail, Event, EventFeedback, \
-    PanelApplicant, PanelApplication
+    PanelApplicant, PanelApplication, GuestGroup
 from uber.utils import add_opt, check, localized_now, validate_model
 from uber.forms import load_forms
 
 
 @all_renderable()
 class Root:
-    def index(self, session, message=''):
+    def index(self, session, message='', department_id=None):
+        apps = session.panel_apps()
+        dept = None
+        if department_id:
+            dept = session.department(department_id)
+            apps = apps.filter(PanelApplication.department == department_id)
+            if not apps.count() and not message:
+                message = f"No panel applications found for the {dept.name} department."
         return {
             'message': message,
-            'apps': session.panel_apps()
+            'apps': session.panel_apps(),
+            'department': dept,
         }
 
-    def app(self, session, id, message='', csrf_token='', explanation=None, **params):
+    def app(self, session, id, message='', **params):
         all_tags = session.query(
             func.string_agg(PanelApplication.tags, literal_column("','"))
         ).all()
 
         app = session.panel_application(id)
+        department = session.department(app.department) if app.department != str(c.PANELS) else None
         forms = load_forms(params, app, ['PanelInfo', 'PanelOtherInfo'])
 
         panelist_form_list = ['PanelistInfo', 'PanelistCredentials']
@@ -45,14 +54,23 @@ class Root:
         
         panelist_forms['new'] = load_forms(params, PanelApplicant(), panelist_form_list,
                                            {form_name: 'new' for form_name in panelist_form_list})
+        
+        guest_groups = session.query(GuestGroup).filter(GuestGroup.group_type != c.MIVS).options(
+            joinedload(GuestGroup.group))
+        guests = [(guest.group.id,
+                  '{} ({})'.format(guest.group.name, guest.group.leader.full_name))
+                for guest in guest_groups if guest.group and guest.group.leader
+        ]
 
         return {
             'message': message,
             'app': session.panel_application(id),
+            'department': department,
             'forms': forms,
             'panelist_forms': panelist_forms,
             'panel_tags': list(set([tag for tag in all_tags[0][0].split(',') if tag != ''])
                                ) if all_tags[0] else '[]',
+            'guests': guests,
         }
 
     @ajax
@@ -112,10 +130,11 @@ class Root:
         return {"success": True}
 
     def edit_panelist(self, session, **params):
+        app = session.panel_application(params.get('app_id'))
         if params.get('id') in [None, '', 'None']:
             panelist = PanelApplicant()
             prefix = 'new'
-            panelist.application = session.panel_application(params.get('application_id'))
+            panelist.applications.append(app)
             session.add(panelist)
         else:
             panelist = session.panel_applicant(params.get('id'))
@@ -130,7 +149,7 @@ class Root:
             session.add(panelist)
 
         raise HTTPRedirect('app?id={}&message={} was successfully {}.',
-                           panelist.application.id, panelist.full_name, 'created' if prefix == 'new' else 'updated')
+                           app.id, panelist.full_name, 'created' if prefix == 'new' else 'updated')
 
     def email_statuses(self, session):
         emails = session.query(AutomatedEmail).filter(AutomatedEmail.ident.in_(
@@ -143,6 +162,40 @@ class Root:
             'attendee': attendee,
             'panels': sorted(attendee.panel_applications, key=lambda app: app.name)
         }
+    
+    def assign_guest(self, session, id, group_id, **params):
+        app = session.panel_application(id)
+        group = session.group(group_id)
+        if not group.guest:
+            raise HTTPRedirect('app?id={}&message={}', id, f'{group.name} is not a guest group.')
+        if not group.leader:
+            raise HTTPRedirect('app?id={}&message={}', id,
+                               f'{group.name} does not have a leader to set as the panel submitter.')
+    
+        if app.submitter:
+            app.submitter.check_if_still_submitter(app)
+
+        panelist_attendee = group.leader
+        if panelist_attendee.panel_applicants:
+            panelist = sorted(panelist_attendee.panel_applicants, key=lambda p: p.submitter)[0]
+        else:
+            panelist = PanelApplicant(
+                attendee_id=panelist_attendee.id,
+                first_name=panelist_attendee.first_name,
+                last_name=panelist_attendee.last_name,
+                email=panelist_attendee.email,
+                cellphone=panelist_attendee.cellphone
+            )
+        
+        panelist.applications.append(app)
+        panelist.submitter = True
+        app.submitter_id = panelist.id
+        session.add(panelist)
+        if panelist_attendee.badge_type != c.GUEST_BADGE:
+            add_opt(panelist_attendee.ribbon_ints, c.PANELIST_RIBBON)
+            session.add(panelist_attendee)
+        
+        raise HTTPRedirect('app?id={}&message={}', id, f"Panel successfully assigned to {group.name}.")
 
     @csrf_protected
     def update_comments(self, session, id, comments):
@@ -154,6 +207,7 @@ class Root:
         app = session.panel_application(params)
         if app.status != c.ACCEPTED and int(status) == c.ACCEPTED:
             app.accepted = datetime.now()
+            app.add_credentials_to_desc()
         app.status = int(status)
         if not app.poc:
             app.poc_id = session.admin_attendee().id
@@ -181,18 +235,25 @@ class Root:
         raise HTTPRedirect('app?id={}&message={}', app.id, 'Tags updated')
 
     @csrf_protected
-    def change_submitter(self, session, applicant_id):
+    def change_submitter(self, session, applicant_id, app_id):
+        app = session.panel_application(app_id)
         panelist = session.panel_applicant(applicant_id)
-        for each_panelist in panelist.application.applicants:
-            each_panelist.submitter = False
+        
+        app.submitter.check_if_still_submitter(app_id)
+        app.submitter_id = panelist.id
         panelist.submitter = True
-        raise HTTPRedirect('app?id={}&message=Point of contact was updated to {}', panelist.app_id, panelist.full_name)
+
+        raise HTTPRedirect('app?id={}&message=Point of contact was updated to {}', app.id, panelist.full_name)
 
     @csrf_protected
-    def remove_submitter(self, session, applicant_id):
+    def remove_panelist(self, session, applicant_id, app_id):
+        app = session.panel_application(app_id)
         panelist = session.panel_applicant(applicant_id)
+        if panelist.submitter:
+            raise HTTPRedirect('app?id={}&message={}',
+                               "This panelist is the submitter for another panel and cannot be removed.")
         session.delete(panelist)
-        raise HTTPRedirect('app?id={}&message=Panelist {} was removed', panelist.app_id, panelist.full_name)
+        raise HTTPRedirect('app?id={}&message=Panelist {} was removed', app.id, panelist.full_name)
 
     def associate(self, session, message='', **params):
         app = session.panel_application(params)
@@ -227,7 +288,7 @@ class Root:
 
         applicants = []
         for pa in session.panel_applicants():
-            if not pa.attendee_id and pa.application.status == c.ACCEPTED:
+            if not pa.attendee_id and pa.accepted_applications:
                 applicants.append([pa, set(possibles[pa.email.lower()] + possibles[pa.first_name, pa.last_name])])
 
         return {'applicants': applicants}
